# Pkg stanza extract file reading