--- a/pkg/stanza/operator/input/file/config.go
+++ b/pkg/stanza/operator/input/file/config.go
@@ -15,13 +15,10 @@
 package file // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/file"
 
 import (
-	"fmt"
-	"time"
-
-	"github.com/bmatcuk/doublestar/v3"
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/internal/fileconsumer"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
@@ -30,45 +27,18 @@ func init() {
 	operator.Register("file_input", func() operator.Builder { return NewConfig("") })
 }
 
-const (
-	defaultMaxLogSize         = 1024 * 1024
-	defaultMaxConcurrentFiles = 1024
-)
-
 // NewConfig creates a new input config with default values
 func NewConfig(operatorID string) *Config {
 	return &Config{
-		InputConfig:             helper.NewInputConfig(operatorID, "file_input"),
-		PollInterval:            helper.Duration{Duration: 200 * time.Millisecond},
-		IncludeFileName:         true,
-		IncludeFilePath:         false,
-		IncludeFileNameResolved: false,
-		IncludeFilePathResolved: false,
-		Splitter:                helper.NewSplitterConfig(),
-		StartAt:                 "end",
-		FingerprintSize:         defaultFingerprintSize,
-		MaxLogSize:              defaultMaxLogSize,
-		MaxConcurrentFiles:      defaultMaxConcurrentFiles,
-		Encoding:                helper.NewEncodingConfig(),
+		InputConfig: helper.NewInputConfig(operatorID, "file_input"),
+		Config:      *fileconsumer.NewConfig(),
 	}
 }
 
 // Config is the configuration of a file input operator
 type Config struct {
-	helper.InputConfig `mapstructure:",squash" yaml:",inline"`
-	Finder             `mapstructure:",squash" yaml:",inline"`
-
-	PollInterval            helper.Duration       `mapstructure:"poll_interval,omitempty"                  json:"poll_interval,omitempty"                 yaml:"poll_interval,omitempty"`
-	IncludeFileName         bool                  `mapstructure:"include_file_name,omitempty"              json:"include_file_name,omitempty"             yaml:"include_file_name,omitempty"`
-	IncludeFilePath         bool                  `mapstructure:"include_file_path,omitempty"              json:"include_file_path,omitempty"             yaml:"include_file_path,omitempty"`
-	IncludeFileNameResolved bool                  `mapstructure:"include_file_name_resolved,omitempty"     json:"include_file_name_resolved,omitempty"    yaml:"include_file_name_resolved,omitempty"`
-	IncludeFilePathResolved bool                  `mapstructure:"include_file_path_resolved,omitempty"     json:"include_file_path_resolved,omitempty"    yaml:"include_file_path_resolved,omitempty"`
-	StartAt                 string                `mapstructure:"start_at,omitempty"                       json:"start_at,omitempty"                      yaml:"start_at,omitempty"`
-	FingerprintSize         helper.ByteSize       `mapstructure:"fingerprint_size,omitempty"               json:"fingerprint_size,omitempty"              yaml:"fingerprint_size,omitempty"`
-	MaxLogSize              helper.ByteSize       `mapstructure:"max_log_size,omitempty"                   json:"max_log_size,omitempty"                  yaml:"max_log_size,omitempty"`
-	MaxConcurrentFiles      int                   `mapstructure:"max_concurrent_files,omitempty"           json:"max_concurrent_files,omitempty"          yaml:"max_concurrent_files,omitempty"`
-	Encoding                helper.EncodingConfig `mapstructure:",squash,omitempty"                        json:",inline,omitempty"                       yaml:",inline,omitempty"`
-	Splitter                helper.SplitterConfig `mapstructure:",squash,omitempty"                        json:",inline,omitempty"                       yaml:",inline,omitempty"`
+	helper.InputConfig  `mapstructure:",squash" yaml:",inline"`
+	fileconsumer.Config `mapstructure:",squash" yaml:",inline"`
 }
 
 // Build will build a file input operator from the supplied configuration
@@ -76,61 +46,6 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 	inputOperator, err := c.InputConfig.Build(logger)
 	if err != nil {
 		return nil, err
-	}
-
-	if len(c.Include) == 0 {
-		return nil, fmt.Errorf("required argument `include` is empty")
-	}
-
-	// Ensure includes can be parsed as globs
-	for _, include := range c.Include {
-		_, err = doublestar.PathMatch(include, "matchstring")
-		if err != nil {
-			return nil, fmt.Errorf("parse include glob: %w", err)
-		}
-	}
-
-	// Ensure excludes can be parsed as globs
-	for _, exclude := range c.Exclude {
-		_, err = doublestar.PathMatch(exclude, "matchstring")
-		if err != nil {
-			return nil, fmt.Errorf("parse exclude glob: %w", err)
-		}
-	}
-
-	if c.MaxLogSize <= 0 {
-		return nil, fmt.Errorf("`max_log_size` must be positive")
-	}
-
-	if c.MaxConcurrentFiles <= 1 {
-		return nil, fmt.Errorf("`max_concurrent_files` must be greater than 1")
-	}
-
-	if c.FingerprintSize == 0 {
-		c.FingerprintSize = defaultFingerprintSize
-	} else if c.FingerprintSize < minFingerprintSize {
-		return nil, fmt.Errorf("`fingerprint_size` must be at least %d bytes", minFingerprintSize)
-	}
-
-	encoding, err := c.Encoding.Build()
-	if err != nil {
-		return nil, err
-	}
-
-	// Ensure that multiline is buildable
-	_, err = c.Splitter.Build(encoding.Encoding, false, int(c.MaxLogSize))
-	if err != nil {
-		return nil, err
-	}
-
-	var startAtBeginning bool
-	switch c.StartAt {
-	case "beginning":
-		startAtBeginning = true
-	case "end":
-		startAtBeginning = false
-	default:
-		return nil, fmt.Errorf("invalid start_at location '%s'", c.StartAt)
 	}
 
 	fileNameField := entry.NewNilField()
@@ -153,25 +68,18 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		filePathResolvedField = entry.NewAttributeField("log.file.path_resolved")
 	}
 
-	return &Input{
+	input := &Input{
 		InputOperator:         inputOperator,
-		finder:                c.Finder,
-		PollInterval:          c.PollInterval.Raw(),
 		FilePathField:         filePathField,
 		FileNameField:         fileNameField,
 		FilePathResolvedField: filePathResolvedField,
 		FileNameResolvedField: fileNameResolvedField,
-		startAtBeginning:      startAtBeginning,
-		Splitter:              c.Splitter,
-		queuedMatches:         make([]string, 0),
-		encoding:              encoding,
-		firstCheck:            true,
-		cancel:                func() {},
-		knownFiles:            make([]*Reader, 0, 10),
-		roller:                newRoller(),
-		fingerprintSize:       int(c.FingerprintSize),
-		MaxLogSize:            int(c.MaxLogSize),
-		MaxConcurrentFiles:    c.MaxConcurrentFiles,
-		SeenPaths:             make(map[string]struct{}, 100),
-	}, nil
+	}
+
+	input.fileConsumer, err = c.Config.Build(logger, input.emit)
+	if err != nil {
+		return nil, err
+	}
+
+	return input, nil
 }