--- a/pkg/stanza/operator/input/file/rotation_test.go
+++ b/pkg/stanza/operator/input/file/rotation_test.go
@@ -15,7 +15,6 @@
 package file
 
 import (
-	"context"
 	"fmt"
 	"io"
 	"os"
@@ -354,25 +353,21 @@ func TestMoveFile(t *testing.T) {
 	}
 	t.Parallel()
 	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-	operator.persister = testutil.NewMockPersister("test")
 
 	temp1 := openTemp(t, tempDir)
 	writeString(t, temp1, "testlog1\n")
 	temp1.Close()
 
-	operator.poll(context.Background())
+	require.NoError(t, operator.Start(testutil.NewMockPersister("test")))
 	defer func() {
 		require.NoError(t, operator.Stop())
 	}()
 
 	waitForMessage(t, logReceived, "testlog1")
 
-	// Wait until all goroutines are finished before renaming
-	operator.wg.Wait()
 	err := os.Rename(temp1.Name(), fmt.Sprintf("%s.2", temp1.Name()))
 	require.NoError(t, err)
 
-	operator.poll(context.Background())
 	expectNoMessages(t, logReceived)
 }
 
@@ -382,21 +377,17 @@ func TestTrackMovedAwayFiles(t *testing.T) {
 	}
 	t.Parallel()
 	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-	operator.persister = testutil.NewMockPersister("test")
 
 	temp1 := openTemp(t, tempDir)
 	writeString(t, temp1, "testlog1\n")
 	temp1.Close()
 
-	operator.poll(context.Background())
+	require.NoError(t, operator.Start(testutil.NewMockPersister("test")))
 	defer func() {
 		require.NoError(t, operator.Stop())
 	}()
 
 	waitForMessage(t, logReceived, "testlog1")
-
-	// Wait until all goroutines are finished before renaming
-	operator.wg.Wait()
 
 	newDir := fmt.Sprintf("%s%s", tempDir[:len(tempDir)-1], "_new/")
 	err := os.Mkdir(newDir, 0777)
@@ -409,7 +400,6 @@ func TestTrackMovedAwayFiles(t *testing.T) {
 	movedFile, err := os.OpenFile(newFileName, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	require.NoError(t, err)
 	writeString(t, movedFile, "testlog2\n")
-	operator.poll(context.Background())
 
 	waitForMessage(t, logReceived, "testlog2")
 }
@@ -419,12 +409,11 @@ func TestTrackMovedAwayFiles(t *testing.T) {
 func TestTruncateThenWrite(t *testing.T) {
 	t.Parallel()
 	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-	operator.persister = testutil.NewMockPersister("test")
 
 	temp1 := openTemp(t, tempDir)
 	writeString(t, temp1, "testlog1\ntestlog2\n")
 
-	operator.poll(context.Background())
+	require.NoError(t, operator.Start(testutil.NewMockPersister("test")))
 	defer func() {
 		require.NoError(t, operator.Stop())
 	}()
@@ -437,7 +426,6 @@ func TestTruncateThenWrite(t *testing.T) {
 	require.NoError(t, err)
 
 	writeString(t, temp1, "testlog3\n")
-	operator.poll(context.Background())
 	waitForMessage(t, logReceived, "testlog3")
 	expectNoMessages(t, logReceived)
 }
@@ -449,19 +437,17 @@ func TestTruncateThenWrite(t *testing.T) {
 func TestCopyTruncateWriteBoth(t *testing.T) {
 	t.Parallel()
 	operator, logReceived, tempDir := newTestFileOperator(t, nil, nil)
-	operator.persister = testutil.NewMockPersister("test")
 
 	temp1 := openTemp(t, tempDir)
 	writeString(t, temp1, "testlog1\ntestlog2\n")
 
-	operator.poll(context.Background())
+	require.NoError(t, operator.Start(testutil.NewMockPersister("test")))
 	defer func() {
 		require.NoError(t, operator.Stop())
 	}()
 
 	waitForMessage(t, logReceived, "testlog1")
 	waitForMessage(t, logReceived, "testlog2")
-	operator.wg.Wait() // wait for all goroutines to finish
 
 	// Copy the first file to a new file, and add another log
 	temp2 := openTemp(t, tempDir)
@@ -478,7 +464,6 @@ func TestCopyTruncateWriteBoth(t *testing.T) {
 	writeString(t, temp1, "testlog4\n")
 
 	// Expect both messages to come through
-	operator.poll(context.Background())
 	waitForMessages(t, logReceived, []string{"testlog3", "testlog4"})
 }