… pkg/stanza/operator/input/file/file_test.go
… pkg/stanza/operator/input/file/finder.go
… pkg/stanza/operator/input/file/finder_test.go
… pkg/stanza/operator/input/file/fingerprint.go
… pkg/stanza/operator/input/file/fingerprint_test.go
… pkg/stanza/operator/input/file/positional_scanner.go
… pkg/stanza/operator/input/file/reader.go
… pkg/stanza/operator/input/file/roller.go
… pkg/stanza/operator/input/file/roller_other.go
… pkg/stanza/operator/input/file/roller_windows.go