This is a major step towards #10286.

I expect to follow this up with some cleanup PRs, but wanted to leave as much as possible in place so as not to further bloat an already huge PR.
- The new package name and location should be considered temporary. I expect that this package will be exposed and available for other use cases, but would like to at least clean it up before doing this.
- Some variable names are a very poor fit in the new design, but were intentionally left in place to minimize the diff.
- The biggest point of deficiency in this design relates to encodings. I think the file tracking logic should either emit `[]byte` and leave decoding to the caller, or it should manage all decoding. Currently, this is somewhat awkwardly split across the interface. This would be a good topic to address in a followup PR.
- Many tests are duplicated, but I believe there is some value in ensuring that the higher level caller is validating expectations on its own.
- The design document is duplicated. Later, I expect it will reside in the new package only.
- Many of the configuration tests need to be reviewed and updated in the new package (eg those that contain operator IDs)