@djaglowski 2022-06-17
There are some minor memory utilization improvements here. This is somewhat expected because `Reader`'s are no longer maintaining their own decode buffers, but instantiating new ones as necessary. This was prompted by the challenge of cleaning up decode logic, mentioned above, but appears to be a notable positive impact.

`main`
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/file
BenchmarkFileInput/Single-10         	 1238762	       983.8 ns/op	     672 B/op	       9 allocs/op
BenchmarkFileInput/Glob-10           	  295359	      4146 ns/op	    2691 B/op	      36 allocs/op
BenchmarkFileInput/MultiGlob-10      	  284752	      4243 ns/op	    2691 B/op	      36 allocs/op
BenchmarkFileInput/MaxConcurrent-10  	  154560	      6515 ns/op	    2690 B/op	      36 allocs/op
BenchmarkFileInput/FngrPrntLarge-10  	 1242770	       983.3 ns/op	     672 B/op	       9 allocs/op
BenchmarkFileInput/FngrPrntSmall-10  	 1294358	       961.0 ns/op	     672 B/op	       9 allocs/op
PASS
coverage: 72.0% of statements
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/file	54.505s
```

This branch:
```
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/internal/fileconsumer
BenchmarkFileInput/Single-10         	 1298786	       938.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkFileInput/Glob-10           	  244022	      4179 ns/op	       3 B/op	       0 allocs/op
BenchmarkFileInput/MultiGlob-10      	  246051	      4086 ns/op	       3 B/op	       0 allocs/op
BenchmarkFileInput/MaxConcurrent-10  	  134785	      7449 ns/op	       2 B/op	       0 allocs/op
BenchmarkFileInput/FngrPrntLarge-10  	 1329999	       921.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkFileInput/FngrPrntSmall-10  	 1037469	       978.9 ns/op	       0 B/op	       0 allocs/op
PASS
coverage: 73.4% of statements
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/internal/fileconsumer	53.433s
```