@jsirianni 2022-06-17
> @TylerHelmuth, the files are moved & slightly modified from their previous location in `pkg/stanza/operator/input/file`.
> 
> Perhaps @jsirianni could give this a once over?

Running great for me :+1: